requirements.txt
# Telegram Bot
python-telegram-bot[job-queue]==21.10

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Utilities
python-dotenv==1.0.1
greenlet==3.0.3

# Scheduling
apscheduler==3.10.4

# Web Backend (Priority 3)
fastapi==0.109.0
uvicorn[standard]==0.27.0
jinja2==3.1.3
python-multipart==0.0.6
aiofiles==23.2.1